chore: move model out of sdk
