+++ circuits/main_update_state.input.json
@@ -1,1 +1,1 @@
-{"proof_id":2,"public_value":"2","public_owner":"15192979887997845057152203024157100889419825593778546904840145855281996190088","num_input_notes":"1","output_nc_1":"14596037139847385691627146843484413803877260987842541517985265174270933138929","output_nc_2":"14638973545683092655181582933095663641137398303422702342728915611146435107272","data_tree_root":"12677761724035481798635753759912892882385963979980181217868184448339906167525","siblings":[["19778943385211782290410286996991910910383042879572711378066751126663234016105","9028070735772241879306691186790098011835015517545241153444647405817301683092","11954575606818816921012866400107699336228450618518006592440119706693887268506","0","19885456351023803904725665203596529768208754721373345829199371742997865470211","13871334651200641624204486353806737791046199059538436257730501069692529792467","263266965445291798379879802262497978356589124884569652208699660447423065266",0,0,0,0,0,0,0,0,0,0,0,0,0],["21841842852738631938599452054865981188848503934592422785509626003006692726673","16092318996453270957155512979397248749913146255535385850348699083166767550580","7890025723828357355455362486079674545587169853012934932221618763524946281563","0","21560203916548051822274828076204553406336257877394889773196127587586743630983","0","14712211048120142657840198268696186209323783461075139848687031822817546814413",0,0,0,0,0,0,0,0,0,0,0,0,0]],"siblings_ac":["19778943385211782290410286996991910910383042879572711378066751126663234016105","18321713402743850096722356142671327151154831340463540869781406366690802696405","21112149179583867639143254079641231365844039772774700620098891335339710625090","7749101918053595476826873687977043609076619663712566603385472781709448211206","16932604148873947701919611837867844776824404564032235939470764490654988840209","0","0","0","0","0","0","0","5808798268272931466212217219512695973676238229447086612884160658160656760337",0,0,0,0,0,0,0],"asset_id":1,"public_asset_id":1,"alias_hash":"123","input_note_val":["5","0"],"input_note_secret":["15372927681730784543286638844669647497274326642792362601359107337649795717906","0"],"input_note_asset_id":["1","1"],"input_note_owner":[["15068083104205934649543480837376852248491359267450502981674755036693835917874","2040696816059603052998238231571501248771651498833683695961169530468099778962"],["15068083104205934649543480837376852248491359267450502981674755036693835917874","2040696816059603052998238231571501248771651498833683695961169530468099778962"]],"input_note_nullifier":["4327332458499936979663641042694707121628453113066511760052968083906302686700","9889731345309911758861447299955773457036954297182515905143653757571926522898"],"input_note_account_required":["0","0"],"output_note_val":["0","3"],"output_note_secret":["12327599883929078287812100844707159861362898502648888474444077662643154165127","12327599883929078287812100844707159861362898502648888474444077662643154165127"],"output_note_asset_id":["1","1"],"output_note_owner":[["889051278736292297292753339794175271527267418357559169658462682666886857620","20362772477919704444392529882551145492652160623268762095856687336540749128731"],["15068083104205934649543480837376852248491359267450502981674755036693835917874","2040696816059603052998238231571501248771651498833683695961169530468099778962"]],"output_note_nullifier":["20628619486515588536032446902596700078370020418703084867338220662655730198462","15244663473769184966257859882611156949866004851422048540775220921803816287335"],"output_note_account_required":["0","0"],"account_required":false,"account_note_nk":"4585798145318716933389973202665892955907519804637393867075402739026670088434","account_note_npk":["15068083104205934649543480837376852248491359267450502981674755036693835917874","2040696816059603052998238231571501248771651498833683695961169530468099778962"],"account_note_spk":["889051278736292297292753339794175271527267418357559169658462682666886857620","20362772477919704444392529882551145492652160623268762095856687336540749128731"],"signatureR8":["16662843779865551653436132308853328512136364965663908329861751467786792031893","2923862563471055391289842869213799669446317730671598611411197969847284585243"],"signatureS":"400965765515767628969124506283516048161519525308871804816654428407065616815","new_account_note_npk":["0","0"],"new_account_note_spk1":["0","0"],"new_account_note_spk2":["0","0"]}
+{"proof_id":1,"public_value":"10","public_owner":"20803409528390686597841086667104243207638346542581965116496083283645815025795","num_input_notes":"0","output_nc_1":"19457382994828088408705033937951626752470268198339416285286528148451447916181","output_nc_2":"18207070648711432099580512251739008856357889930371283934864076624106747145582","data_tree_root":"4483476425698921923888070293619352992537586414614148774365727897591320015120","siblings":[["17567611789758876561513565510903156219513783781124217291527914446858046571992","11206175460100387262411895429131700829003002291847572010715501133972450222356",0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],["3286565227609673646221992011170375962675773465772137208234527822622768211054",0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]],"siblings_ac":["17567611789758876561513565510903156219513783781124217291527914446858046571992","11206175460100387262411895429131700829003002291847572010715501133972450222356",0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],"asset_id":1,"public_asset_id":1,"alias_hash":"21106690491256420095915831706181313561710745369390204953383467951642800824149","input_note_val":["0","0"],"input_note_secret":["0","0"],"input_note_asset_id":["1","1"],"input_note_owner":[["1508404322884523863061850177648188865407443616422762673697724360683563115204","16515811321707783139143639069964779422691349706657981043294415585904997069992"],["1508404322884523863061850177648188865407443616422762673697724360683563115204","16515811321707783139143639069964779422691349706657981043294415585904997069992"]],"input_note_nullifier":["11276206492787332549510060197113924068401192354026817286836041057997683020737","2901290505065624903404314018486663476543044420211115040680024663762551666999"],"input_note_account_required":["0","0"],"output_note_val":["10","0"],"output_note_secret":["14599413254934758096020545899784814156464394203838224647984036696405881515461","14599413254934758096020545899784814156464394203838224647984036696405881515461"],"output_note_asset_id":["1","1"],"output_note_owner":[["1508404322884523863061850177648188865407443616422762673697724360683563115204","16515811321707783139143639069964779422691349706657981043294415585904997069992"],["1508404322884523863061850177648188865407443616422762673697724360683563115204","16515811321707783139143639069964779422691349706657981043294415585904997069992"]],"output_note_nullifier":["18466683771449098687963281403685916834406449691199403601712142004817270619051","19881037944154990517811902984312484580493525610868928608734480485260600679277"],"output_note_account_required":["0","0"],"account_required":false,"account_note_nk":"3899344576546501402780394572041678009130745634382488157924715101585674676175","account_note_npk":["1508404322884523863061850177648188865407443616422762673697724360683563115204","16515811321707783139143639069964779422691349706657981043294415585904997069992"],"account_note_spk":["20803409528390686597841086667104243207638346542581965116496083283645815025795","17091158337967369684135630565782979174878809214565202750682326945954892583734"],"signatureR8":["10879894993349779707579309375692060323616782315808035437684028870811003629692","9246114001582445102562217575885192700988071395474193103582812656236966913580"],"signatureS":"376704356458846143110323334037091169362329556575411747620054239608394297142","new_account_note_npk":["0","0"],"new_account_note_spk1":["0","0"],"new_account_note_spk2":["0","0"]}
