+++ circuits/main_withdraw.input.json
@@ -1,1 +1,1 @@
-{"enabled":1,"R8x":"7521711703078129114294509701378571364456404766481662390871780629538542027185","R8y":"17720148169625746400663542145140387861360653906251972796264886638614176441908","S":"1555790952131803508302247779404634770931741780003560988817573988301810531621","Ax":"21033607372082967370241374913179996986666042889747562125232994613648413442739","Ay":"6687258640429679799689964574036709272478698642520002464129951093950855145421","M":"1979475358490882782695234604362398132934050455360496620085373760138828661113"}
+{"enabled":1,"R8x":"6559052220100888891224284827212785738814637782496117018809379249015168045188","R8y":"3166736329535828063212969300801834233849123916597787952865105209612678553788","S":"1381304026404232124401130511590765322974053223860252089487503101735023415360","Ax":"21798931248873568419354832279394015192266255554393305594854879173831472525940","Ay":"3634186412603978995064046842950718124874269315448335864734242543201107295800","M":"1979475358490882782695234604362398132934050455360496620085373760138828661113"}
