+++ src/utils.ts
@@ -1,8 +1,13 @@
 import { Buffer } from "buffer";
 import { BigNumberish } from "ethers";
-import { N_LEVEL } from "./state_tree";
+import { N_LEVEL } from "./state_tree_circuit";
 import { ethers } from "ethers";
 import consola from "consola";
+import { randomBytes as _randomBytes } from "crypto";
+
+export function index() {
+    return BigInt("0x" + _randomBytes(31).toString("hex"))
+}
 
 export function arrayChunk(array: Array<number>, chunkSize: number): any {
   return Array(Math.ceil(array.length / chunkSize)).map((_, index) => index * chunkSize)
