+++ tasks/index.ts
@@ -1,1 +1,1 @@
-import './Account'
+import './account'
