+++ test/end_to_end.test.ts
@@ -8,8 +8,8 @@ import { ethers } from "ethers";
 import { uint8Array2Bigint, signEOASignature, prepareJson } from "../src/utils";
 import { expect, assert } from "chai";
 import { StateTree } from "../src/state_tree";
-import { NoteState, NoteModel as DBNote } from "../server/note";
-import { Note } from "../src/note";
+import { NoteModel as DBNote } from "../server/note";
+import { Note, NoteState } from "../src/note";
 import { aliasHashDigest, compress as accountCompress, EigenAddress, SigningKey } from "../src/account";
 import { JoinSplitCircuit } from "../src/join_split";
 import { AccountCircuit } from "../src/account";
