+++ test/note.test.ts
@@ -3,6 +3,7 @@ import { Note } from "../src/note";
 import { assert, expect } from "chai";
 import { ethers } from "ethers";
 import { StateTree } from "../src/state_tree";
+import { index } from "../src/utils";
 
 const { buildEddsa, buildBabyjub } = require("circomlibjs");
 import { SigningKey } from "../src/account";
@@ -24,7 +25,7 @@ describe("Test NoteCompressor", function () {
 
     it("Note compress test", async () => {
         let key = await (new SigningKey()).newKey(undefined);
-        let note = new Note(1n, 2n, key.pubKey, 4, 5n, true, StateTree.index);
+        let note = new Note(1n, 2n, key.pubKey, 4, 5n, true, index());
         let wtns = await utils.executeCircuit(circuit, note.toCircuitInput(babyJub));
         await circuit.assertOut(wtns, { out: await note.compress(babyJub) });
     })
